Filter out null specs ending up in split_configuration
This is awkward because we're not exactly sure how a null value is ending up in a list of files in split_configuration.json. We're updating the code to handle it better than it currently does - blows up.
We're still investigating a possible root cause.

Gem version updated from 2.7.0 to 2.7.0.1